+++ shared/src/iosMain/kotlin/eu/heha/conifer/IosPreferencesInitializer.kt
@@ -3,12 +3,11 @@ package eu.heha.conifer
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
-import eu.heha.conifer.prefs.SyncPrefs
 import okio.Path.Companion.toPath
 
-object IosSyncPrefsInitializer : SyncPrefsInitializer {
-    override fun createSyncPrefsStore(): DataStore<Preferences> =
+object IosPreferencesInitializer : PreferencesInitializer {
+    override fun createStore(fileName: String): DataStore<Preferences> =
         PreferenceDataStoreFactory.createWithPath {
-            (iosDocumentDirectory() + "/" + SyncPrefs.STORE_FILE_NAME).toPath()
+            (iosDocumentDirectory() + "/" + fileName).toPath()
         }
 }