+++ shared/src/jvmMain/kotlin/eu/heha/conifer/JvmPreferencesInitializer.kt
@@ -3,13 +3,12 @@ package eu.heha.conifer
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
-import eu.heha.conifer.prefs.SyncPrefs
 import okio.Path.Companion.toOkioPath
 import java.io.File
 
-object JvmSyncPrefsInitializer : SyncPrefsInitializer {
-    override fun createSyncPrefsStore(): DataStore<Preferences> =
+object JvmPreferencesInitializer : PreferencesInitializer {
+    override fun createStore(fileName: String): DataStore<Preferences> =
         PreferenceDataStoreFactory.createWithPath {
-            File(jvmDataFolder(), SyncPrefs.STORE_FILE_NAME).toOkioPath()
+            File(jvmDataFolder(), fileName).toOkioPath()
         }
 }